pyproject.toml
[project]
name = "model-based_rl_main"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "james_dell", email = "james.rudd.jones@gmail.com"},
]
dependencies = [
    "absl-py>=2.1.0",
    "ml-collections>=1.0.0",
    "wandb>=0.19.8",
    "matplotlib>=3.10.1",
    "neatplot>=0.1.0",
    "gymnax>=0.0.8",
    "jupyter>=1.1.1",
    "notebook>=7.3.3",
    "tensorflow-probability>=0.25.0",
    "jax[cuda]==0.6.0",
    "gpjax>=0.12.2",
    "bifurcagym @ git+ssh://git@github.com/JamesRudd-Jones/BifurcaGym.git@main",
]
#dependencies = [
#    "absl-py>=2.1.0",
#    "ml-collections>=1.0.0",
#    "wandb>=0.19.8",
#    "matplotlib>=3.10.1",
#    "neatplot>=0.1.0",
#    "gymnax>=0.0.8",
#    "jupyter>=1.1.1",
#    "notebook>=7.3.3",
#    "gpjax>=0.11.0",
#    "tensorflow-probability>=0.25.0",
#    "jax[cuda]==0.6.0",
#]

#dependencies = [
#    "absl-py>=2.1.0",
#    "ml-collections>=1.0.0",
#    "wandb>=0.19.8",
#    "matplotlib>=3.10.1",
#    "neatplot>=0.1.0",
#    "gymnax>=0.0.8",
#    "gpjax>=0.9.4",
##    "jax[cuda12_pip]<0.4.20",  # TODO this is due to some weird segmentation errors caused by jax 0.4.27???????
#    "scipy<=1.12.0,>=1.6.0",  # for the low jax package
#    "jupyter>=1.1.1",
#    "notebook>=7.3.3",
#]

requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[[tool.pdm.source]]
 name = "jax_cuda"
 url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
 verify_ssl = true
 type = "find_links"

[tool.pdm.scripts]
test.cmd = "pytest -n auto --maxprocesses 8 --dist loadfile"
# test.env = { JAX_PLATFORM_NAME = "cpu", JAX_DISABLE_MOST_OPTIMIZATIONS = "1" }
lint.shell = "ruff . && black --check ."
fmt.shell = "ruff --fix . && black ."

[tool.pdm.dev-dependencies]
dev = [
    "lxm3>=0.4.0",
    "ruff>=0.1.11",
    "black>=23.12.1",
    "pytest>=7.4.4",
    "pytest-xdist>=3.5.0",
    "pre-commit>=3.6.0",
]

# Ruff configuration
# https://docs.astral.sh/ruff/configuration/
[tool.ruff]
line-length = 120
select = [
    "E", # pycodestyle
    "F", # pyflakes
    "I", # isort
]
ignore = ["E731"]

[tool.ruff.extend-per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
force-single-line = true
force-sort-within-sections = true
single-line-exclusions = ["typing"]

# Black configuration
# https://black.readthedocs.io/en/stable/usage_and_configuration/the_basics.html
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.github
  | \.venv
  | docs
  | third_party
  | scratch
)/
'''